Fix type casting for index comparison in residualbased_block_builder_and_solver

+++ kratos/solving_strategies/builder_and_solvers/residualbased_block_builder_and_solver.h
@@ -893,7 +893,7 @@ public:
             if (k_factor == 0.0) {
                 // Zero out the whole row, except the diagonal
                 for (std::size_t j = col_begin; j < col_end; ++j)
-                    if (Acol_indices[j] != Index )
+                    if (Acol_indices[j] != static_cast<long int>(Index) )
                         Avalues[j] = 0.0;
 
                 // Zero out the RHS
